[tests] Don't use symlinks, it doesn't work on Windows :/
Instead copy each file 4 times, and add a notice to remember to update the
other identical files if any of them are modified.
