--- tests/dotnet/BindingWithDefaultCompileInclude/ApiDefinition.cs
@@ -1,7 +1,0 @@
-using Foundation;
-
-namespace MyApiDefinition {
-	[BaseType (typeof (NSObject))]
-	interface MyNativeClass {
-	}
-}
