+++ tests/dotnet/BindingWithDefaultCompileInclude/MacCatalyst/MyClass.cs
@@ -1,1 +1,9 @@
-../MyClass.cs
+/* If this file is modified, remember to copy the changes the corresponding file for all the other platforms */
+using System;
+namespace MyClassLibrary {
+	public class MyClass {
+		public MyClass ()
+		{
+		}
+	}
+}
