--- tests/dotnet/BindingWithDefaultCompileInclude/MyClass.cs
@@ -1,8 +1,0 @@
-using System;
-namespace MyClassLibrary {
-	public class MyClass {
-		public MyClass ()
-		{
-		}
-	}
-}
