--- tests/dotnet/BindingWithDefaultCompileInclude/StructsAndEnums.cs
@@ -1,6 +1,0 @@
-namespace MyClassLibrary {
-	public struct MyStruct {
-		public int A;
-		public int B;
-	}
-}
