+++ tests/dotnet/BindingWithDefaultCompileInclude/iOS/ApiDefinition.cs
@@ -1,1 +1,8 @@
-../ApiDefinition.cs
+/* If this file is modified, remember to copy the changes the corresponding file for all the other platforms */
+using Foundation;
+
+namespace MyApiDefinition {
+	[BaseType (typeof (NSObject))]
+	interface MyNativeClass {
+	}
+}
