+++ tests/dotnet/BindingWithDefaultCompileInclude/iOS/StructsAndEnums.cs
@@ -1,1 +1,7 @@
-../StructsAndEnums.cs
+/* If this file is modified, remember to copy the changes the corresponding file for all the other platforms */
+namespace MyClassLibrary {
+	public struct MyStruct {
+		public int A;
+		public int B;
+	}
+}
